Module rome_js_factory::make

Structs

JsArrowFunctionExpressionBuilder

JsBreakStatementBuilder

JsCallExpressionBuilder

JsCatchClauseBuilder

JsCatchDeclarationBuilder

JsClassDeclarationBuilder

JsClassExportDefaultDeclarationBuilder

JsClassExpressionBuilder

JsComputedMemberExpressionBuilder

JsContinueStatementBuilder

JsDebuggerStatementBuilder

JsDirectiveBuilder

JsDoWhileStatementBuilder

JsExportDefaultDeclarationClauseBuilder

JsExportDefaultExpressionClauseBuilder

JsExportFromClauseBuilder

JsExportNamedClauseBuilder

JsExportNamedFromClauseBuilder

JsExportNamedFromSpecifierBuilder

JsExportNamedShorthandSpecifierBuilder

JsExportNamedSpecifierBuilder

JsExpressionStatementBuilder

JsExtendsClauseBuilder

JsForOfStatementBuilder

JsForStatementBuilder

JsFormalParameterBuilder

JsFunctionDeclarationBuilder

JsFunctionExportDefaultDeclarationBuilder

JsFunctionExpressionBuilder

JsGetterClassMemberBuilder

JsGetterObjectMemberBuilder

JsIfStatementBuilder

JsImportBareClauseBuilder

JsImportBuilder

JsImportDefaultClauseBuilder

JsImportNamedClauseBuilder

JsImportNamespaceClauseBuilder

JsMethodClassMemberBuilder

JsMethodObjectMemberBuilder

JsModuleBuilder

JsNamedImportSpecifierBuilder

JsNewExpressionBuilder

JsObjectAssignmentPatternPropertyBuilder

JsObjectAssignmentPatternShorthandPropertyBuilder

JsObjectBindingPatternPropertyBuilder

JsObjectBindingPatternShorthandPropertyBuilder

JsPropertyClassMemberBuilder

JsRestParameterBuilder

JsReturnStatementBuilder

JsScriptBuilder

JsShorthandNamedImportSpecifierBuilder

JsTemplateBuilder

JsThrowStatementBuilder

JsTryFinallyStatementBuilder

JsVariableDeclarationClauseBuilder

JsVariableDeclaratorBuilder

JsVariableStatementBuilder

JsYieldArgumentBuilder

JsYieldExpressionBuilder

JsxAttributeBuilder

JsxExpressionChildBuilder

JsxOpeningElementBuilder

JsxSelfClosingElementBuilder

TsAssertsReturnTypeBuilder

TsBigIntLiteralTypeBuilder

TsCallSignatureTypeMemberBuilder

TsConstructSignatureTypeMemberBuilder

TsConstructorSignatureClassMemberBuilder

TsConstructorTypeBuilder

TsDeclareFunctionDeclarationBuilder

TsEnumDeclarationBuilder

TsEnumMemberBuilder

TsExportAsNamespaceClauseBuilder

TsExportAssignmentClauseBuilder

TsExternalModuleDeclarationBuilder

TsFunctionTypeBuilder

TsGetterSignatureClassMemberBuilder

TsGetterSignatureTypeMemberBuilder

TsImportEqualsDeclarationBuilder

TsImportTypeBuilder

TsIndexSignatureClassMemberBuilder

TsIndexSignatureTypeMemberBuilder

TsInterfaceDeclarationBuilder

TsIntersectionTypeBuilder

TsMappedTypeBuilder

TsMappedTypeOptionalModifierClauseBuilder

TsMappedTypeReadonlyModifierClauseBuilder

TsMethodSignatureClassMemberBuilder

TsMethodSignatureTypeMemberBuilder

TsNameWithTypeArgumentsBuilder

TsNamedTupleTypeElementBuilder

TsNumberLiteralTypeBuilder

TsOptionalPropertyAnnotationBuilder

TsPropertySignatureClassMemberBuilder

TsPropertySignatureTypeMemberBuilder

TsReferenceTypeBuilder

TsSetterSignatureClassMemberBuilder

TsSetterSignatureTypeMemberBuilder

TsThisParameterBuilder

TsTypeAliasDeclarationBuilder

TsTypeParameterBuilder

TsTypeofTypeBuilder

TsUnionTypeBuilder

Functions

eof

ident

Create a new identifier token with no attached trivia

import_meta

js_array_assignment_pattern

js_array_assignment_pattern_element_list

js_array_assignment_pattern_rest_element

js_array_binding_pattern

js_array_binding_pattern_element_list

js_array_binding_pattern_rest_element

js_array_element_list

js_array_expression

js_array_hole

js_arrow_function_expression

js_assignment_expression

js_assignment_with_default

js_await_expression

js_big_int_literal_expression

js_binary_expression

js_binding_pattern_with_default

js_block_statement

js_boolean_literal_expression

js_break_statement

js_call_argument_list

js_call_arguments

js_call_expression

js_case_clause

js_catch_clause

js_catch_declaration

js_class_declaration

js_class_export_default_declaration

js_class_expression

js_class_member_list

js_computed_member_assignment

js_computed_member_expression

js_computed_member_name

js_conditional_expression

js_constructor_class_member

js_constructor_modifier_list

js_constructor_parameter_list

js_constructor_parameters

js_continue_statement

js_debugger_statement

js_default_clause

js_default_import_specifier

js_directive

js_directive_list

js_do_while_statement

js_else_clause

js_empty_class_member

js_empty_statement

js_export

js_export_as_clause

js_export_default_declaration_clause

js_export_default_expression_clause

js_export_from_clause

js_export_named_clause

js_export_named_from_clause

js_export_named_from_specifier

js_export_named_from_specifier_list

js_export_named_shorthand_specifier

js_export_named_specifier

js_export_named_specifier_list

js_expression_snipped

js_expression_statement

js_extends_clause

js_finally_clause

js_for_in_statement

js_for_of_statement

js_for_statement

js_for_variable_declaration

js_formal_parameter

js_function_body

js_function_declaration

js_function_export_default_declaration

js_function_expression

js_getter_class_member

js_getter_object_member

js_identifier_assignment

js_identifier_binding

js_identifier_expression

js_if_statement

js_import

js_import_assertion

js_import_assertion_entry

js_import_assertion_entry_list

js_import_bare_clause

js_import_call_expression

js_import_default_clause

js_import_named_clause

js_import_namespace_clause

js_in_expression

js_initializer_clause

js_instanceof_expression

js_labeled_statement

js_literal_export_name

js_literal_member_name

js_logical_expression

js_method_class_member

js_method_modifier_list

js_method_object_member

js_module

js_module_item_list

js_module_source

js_name

js_named_import_specifier

js_named_import_specifier_list

js_named_import_specifiers

js_namespace_import_specifier

js_new_expression

js_null_literal_expression

js_number_literal

Create a new string literal token with no attached trivia

js_number_literal_expression

js_object_assignment_pattern

js_object_assignment_pattern_property

js_object_assignment_pattern_property_list

js_object_assignment_pattern_rest

js_object_assignment_pattern_shorthand_property

js_object_binding_pattern

js_object_binding_pattern_property

js_object_binding_pattern_property_list

js_object_binding_pattern_rest

js_object_binding_pattern_shorthand_property

js_object_expression

js_object_member_list

js_parameter_list

js_parameters

js_parenthesized_assignment

js_parenthesized_expression

js_post_update_expression

js_pre_update_expression

js_private_class_member_name

js_private_name

js_property_class_member

js_property_modifier_list

js_property_object_member

js_reference_identifier

js_regex_literal_expression

js_rest_parameter

js_return_statement

js_script

js_sequence_expression

js_setter_class_member

js_setter_object_member

js_shorthand_named_import_specifier

js_shorthand_property_object_member

js_spread

js_statement_list

js_static_initialization_block_class_member

js_static_member_assignment

js_static_member_expression

js_static_modifier

js_string_literal

Create a new string literal token with no attached trivia

js_string_literal_expression

js_super_expression

js_switch_case_list

js_switch_statement

js_template

js_template_chunk_element

js_template_element

js_template_element_list

js_this_expression

js_throw_statement

js_try_finally_statement

js_try_statement

js_unary_expression

js_unknown

js_unknown_assignment

js_unknown_binding

js_unknown_expression

js_unknown_import_assertion_entry

js_unknown_member

js_unknown_named_import_specifier

js_unknown_parameter

js_unknown_statement

js_variable_declaration

js_variable_declaration_clause

js_variable_declarator

js_variable_declarator_list

js_variable_statement

js_while_statement

js_with_statement

js_yield_argument

js_yield_expression

jsx_attribute

jsx_attribute_initializer_clause

jsx_attribute_list

jsx_child_list

jsx_closing_element

jsx_closing_fragment

jsx_element

jsx_expression_attribute_value

jsx_expression_child

jsx_fragment

jsx_ident

Create a new identifier token with no attached trivia

jsx_member_name

jsx_name

jsx_namespace_name

jsx_opening_element

jsx_opening_fragment

jsx_reference_identifier

jsx_self_closing_element

jsx_spread_attribute

jsx_spread_child

jsx_string

jsx_tag_expression

jsx_text

new_target

token

Create a new token with the specified syntax kind and no attached trivia

token_decorated_with_space

Create a new token with the specified syntax kind, and a whitespace trivia piece on both the leading and trailing positions

ts_abstract_modifier

ts_accessibility_modifier

ts_any_type

ts_array_type

ts_as_assignment

ts_as_expression

ts_asserts_condition

ts_asserts_return_type

ts_big_int_literal_type

ts_bigint_type

ts_boolean_literal_type

ts_boolean_type

ts_call_signature_type_member

ts_conditional_type

ts_construct_signature_type_member

ts_constructor_signature_class_member

ts_constructor_type

ts_declare_function_declaration

ts_declare_modifier

ts_declare_statement

ts_default_type_clause

ts_definite_property_annotation

ts_definite_variable_annotation

ts_empty_external_module_declaration_body

ts_enum_declaration

ts_enum_member

ts_enum_member_list

ts_export_as_namespace_clause

ts_export_assignment_clause

ts_export_declare_clause

ts_extends_clause

ts_external_module_declaration

ts_external_module_reference

ts_function_type

ts_getter_signature_class_member

ts_getter_signature_type_member

ts_global_declaration

ts_identifier_binding

ts_implements_clause

ts_import_equals_declaration

ts_import_type

ts_import_type_qualifier

ts_index_signature_class_member

ts_index_signature_modifier_list

ts_index_signature_parameter

ts_index_signature_type_member

ts_indexed_access_type

ts_infer_type

ts_instantiation_expression

ts_interface_declaration

ts_intersection_type

ts_intersection_type_element_list

ts_mapped_type

ts_mapped_type_as_clause

ts_mapped_type_optional_modifier_clause

ts_mapped_type_readonly_modifier_clause

ts_method_signature_class_member

ts_method_signature_modifier_list

ts_method_signature_type_member

ts_module_block

ts_module_declaration

ts_name_with_type_arguments

ts_named_tuple_type_element

ts_never_type

ts_non_null_assertion_assignment

ts_non_null_assertion_expression

ts_non_primitive_type

ts_null_literal_type

ts_number_literal_type

ts_number_type

ts_object_type

ts_optional_property_annotation

ts_optional_tuple_type_element

ts_override_modifier

ts_parenthesized_type

ts_predicate_return_type

ts_property_parameter

ts_property_parameter_modifier_list

ts_property_signature_class_member

ts_property_signature_modifier_list

ts_property_signature_type_member

ts_qualified_module_name

ts_qualified_name

ts_readonly_modifier

ts_reference_type

ts_rest_tuple_type_element

ts_return_type_annotation

ts_setter_signature_class_member

ts_setter_signature_type_member

ts_string_literal_type

ts_string_type

ts_symbol_type

ts_template_chunk_element

ts_template_element

ts_template_element_list

ts_template_literal_type

ts_this_parameter

ts_this_type

ts_tuple_type

ts_tuple_type_element_list

ts_type_alias_declaration

ts_type_annotation

ts_type_argument_list

ts_type_arguments

ts_type_assertion_assignment

ts_type_assertion_expression

ts_type_constraint_clause

ts_type_list

ts_type_member_list

ts_type_operator_type

ts_type_parameter

ts_type_parameter_list

ts_type_parameter_name

ts_type_parameters

ts_typeof_type

ts_undefined_type

ts_union_type

ts_union_type_variant_list

ts_unknown_type

ts_void_type